requirements.txt
absl-py==2.0.0
alembic==1.7.6.dev0
altair==5.2.0
annotated-types==0.7.0
anyio==4.0.0
apturl==0.5.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
asttokens==2.4.0
astunparse==1.6.3
async-lru==2.0.4
attrs==23.1.0
axios==0.4.0
Babel==2.12.1
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.7.0
Brlapi==0.8.3
cachetools==5.3.2
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.4
comm==0.1.4
command-not-found==0.3
contourpy==1.1.1
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dash==2.15.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
DateTime==5.4
dbus-python==1.2.18
debugpy==1.8.0
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.7
distro==1.7.0
distro-info===1.1build1
Django==4.2.5
django-cors-headers==4.3.0
django-phonenumber-field==7.2.0
djangorestframework==3.14.0
dnspython==2.7.0
docopt==0.6.2
dotenv==0.9.9
duplicity==0.8.21
email_validator==2.2.0
env==0.1.0
exceptiongroup==1.1.3
executing==1.2.0
fastapi==0.115.8
fasteners==0.14.1
fastjsonschema==2.18.0
filelock==3.12.3
Flask==3.0.0
flatbuffers==23.5.26
fonttools==4.42.1
fqdn==1.5.1
future==0.18.2
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.60.0
gyp==0.1
h11==0.14.0
h5py==3.10.0
heapq-max==0.21
httpcore==1.0.4
httplib2==0.20.2
httpx==0.26.0
idna==3.3
importlib-metadata==4.6.4
ipykernel==6.25.2
ipython==8.15.0
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.0
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.2
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.6
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
keras==2.15.0
keyring==23.5.0
kiwi==9.25.19
kiwisolver==1.4.5
lab==7.4
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==16.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.9.3
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mediapipe==0.10.9
mistune==3.0.1
ml-dtypes==0.2.0
monotonic==1.6
more-itertools==8.10.0
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.8
netifaces==0.11.0
networkx==3.2.1
notebook==7.0.4
notebook_shim==0.2.3
npm==0.1.1
numpy==1.26.0
nvidia-cublas-cu12==12.2.5.6
nvidia-cuda-cupti-cu12==12.2.142
nvidia-cuda-nvcc-cu12==12.2.140
nvidia-cuda-nvrtc-cu12==12.2.140
nvidia-cuda-runtime-cu12==12.2.140
nvidia-cudnn-cu12==8.9.4.25
nvidia-cufft-cu12==11.0.8.103
nvidia-curand-cu12==10.3.3.141
nvidia-cusolver-cu12==11.5.2.141
nvidia-cusparse-cu12==12.1.2.141
nvidia-nccl-cu12==2.16.5
nvidia-nvjitlink-cu12==12.2.140
oauthlib==3.2.0
olefile==0.46
opencv-contrib-python==4.8.1.78
opencv-python==4.8.1.78
opt-einsum==3.3.0
optional-django==0.1.0
overrides==7.4.0
packaging==23.1
pandas==2.1.4
pandocfilters==1.5.0
paramiko==2.9.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pipenv==2023.9.8
platformdirs==3.10.0
plotly==5.19.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==3.20.3
psutil==5.9.5
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.8.1b0
Pygments==2.16.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyMySQL==1.1.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-dotenv==1.0.1
python-json-logger==2.0.7
python-telegram-bot==20.8
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.0
referencing==0.30.2
regex==2023.12.25
reportlab==3.6.8
requests==2.31.0
requests-oauthlib==1.3.1
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.10.3
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.0
SecretStorage==3.3.1
Send2Trash==1.8.2
simplejson==3.19.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sounddevice==0.4.6
soupsieve==2.5
SQLAlchemy==1.4.31
sqlparse==0.4.4
stack-data==0.6.2
starlette==0.45.3
streamlit==1.29.0
systemd-python==234
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0.post1
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.35.0
termcolor==2.4.0
terminado==0.17.1
threadpoolctl==3.2.0
tinycss2==1.2.1
tk==0.1.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.3.3
traitlets==5.10.1
txt2tags==3.8
typing_extensions==4.12.2
tzdata==2023.3
tzlocal==5.2
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
unicorn==2.1.2
uri-template==1.3.0
urllib3==1.26.5
usb-creator==0.3.7
uvicorn==0.34.0
validators==0.22.0
virtualenv==20.24.5
wadllib==1.3.6
watchdog==3.0.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.3
Werkzeug==3.0.1
widgetsnbextension==4.0.9
wrapt==1.14.1
xdg==5
xgboost==2.0.3
xkit==0.0.0
zipp==1.0.0
zope.interface==6.2
